requirements.txt
Python 3.7.4
torch.__version__ = 1.4.0
torchvision.__version__ = 0.5.0
skimage.__version__ = 0.15.0